Adding new module hold items.
Adding four new hold items to the database in the form of the Charge Module, Target Module, Growth Module, and Fortune Module.  Each of these chips have vastly different effects and their value in battle cannot be understated.  The Charge Module allows instant-use of all charge attacks, the Target Module extends the range of all abilities, the Growth Module doubles the experience and stat bonuses gained in battle, and the Fortune Module boosts the likelihood of item drops and critical hits.