Added hold items to robot battle canvas, updated ability compatibility, tweaked critical hits.
Updated the robot battle canvas function to display hold items if applicable.  Hold item appears under the rest of the stats in an appropriately coloured container with a hover title.  Also updated compatibility settings and defined a global list of moves that can be equipped by anyone.  Also made a major change to critical hits and how they are calculated based on level and turn rather than being randomized.