Updated all abilities to support new module items.
Added support for the Target and Charge Modules to all relevant abilities, allowing a quick-charge or extended range when applicable.  Also updated a few existing abilities with new functionality or tweaked damage.  Special mentions for Super Arm which no longer takes on the type of the current field, Time Arrow and Time Slow which have traded certain effects with each other, and the Fire Chaser which works inversely to how it did before (more damage on faster targets).