Debug mode tweaks, script and stylesheet optimizations.
Removed all instances of the legacy debug mode and replaced with something entirely new.  Debug mode is now a toggleable switch in battle that forces the game to show all damage and recovery calculations in great detail.  Debug text may be added elsewhere in the future but right now it's only damage/recovery triggers.  All memory and code optimizations will now be done using webgrind, so the old debug system was no longer necessary.  Also optimized the main stylesheet in the process.