Updated elemental item descriptions.
Revised the descriptions for shards, cores, and stars to be more in line with their current mechanics.  Seeing as cores as no longer throwable that text has been removed and replaced with the holding mechanic.  The text about dark elements in the star descriptions was also removed as it has less to do with them and more to do with the prototype's environment.