Updated appearance conditions for starforce and darkness battles.
Tweaked the single and double battle generator functions with new conditions for the appearance of starforce and/or darkness battles.  Previously you would need to have completed the prototype and the battle and unlocked all robots for a field to generate a star, but now you only need to complete the prototype.  Darkness battles are still difficult to find, however, and still require completing the battle, completing the prototype, and unlocking all masters in that stage.  Starforce should be easy to hunt - darkness battles are just for fun right now.