Adding MySQL object indexes to repository for backup.
Now that player and type data are stored in the database, all major object types are now completely database-driven.  As such I thought it might be a good time to add their MySQL database definitions and data dumps to the repository for backup.  This commit contain the structure and (current) data for players, robots, abilities, items, fields, and types.  Other database tables may be added in future commits but these ones are most relevant to the current update.