Updated damage and recovery triggers to separate starforce calculations from field multipliers, starforce mechanic updated.
Field multipliers and starforce boosts are now calculated separately and can be toggled independently on a per-ability basis.  The Rain Flush ability has been updated to ignore starforce and damage the user (or heal the user in the case of Toad Man) to compensate for the nerf.  Lastly, starforce calculations have been improved when it comes to the defending robot's resistance.  Collecting starforce does not necessarily increase your elemental defense but it does increase your resistance to other player's starforce boosts, nullifying their boost entirely if you have enough of your own.