Recoded admin update functionality, removed hard-coded login updates, created patch notes system.
Removed the legacy file updates script and replaced it with a more flexible and automated system that not only allows for named-patches to be applied and tracked much, much easier but also sends specialized system messages to each user with a breakdown of the update or patch's tweaks applied to their save file.  Very cool stuff.  Also updated community pages to detect system threads and adjust styles/layout.