Updated patches script with new entry for stat_max_limit_update / Balancing Max Limits for Robot Stats
Robot stats are no longer capped at a max of 9999 and are instead based on each robot's base values.  Each stat can now only be boosted to maximum of 5x its base value at level 100.  Any overflow of stats that are now considered too high will be converted to zenny for the player.