Updated stat capsules with new long-term effects, more efficient coding.
Added a new item function to the parent ability class for stat booster items and revised the functionality so that the effects are long-term.  This will make it much, much easier to grind your robots stats to the max level without too much effort.  To balance out their new usefulness, the prices of these items have gone up in the shop considerably.