Updated ability class with new template function plus placeholders for more.
Created the first of several ability template functions (much like the item functions) so it's easy to create multiple abilities using the same basic structure.  This one is for triggering a forward attack on the enemy with an optional additional effect.  Ice Slash is the first ability to be converted to using the new system.