Updated Rain Flush so it doesn't suck anymore.
It is no longer percent based and is instead equal to the user's current level.  Is it affected by starforce, core boost, and even field multipliers.  And best of all, it no longer damages any robots on your own team.  When Rain Flush is activated, your own benched robots cover themselves in a special shield that resists all effects of the attack.