Removed legacy Stat Blaze and Stat Burn abilities from the game.
There are enough Neutral type stat booster abilities as it is and with items available for most of the game there's no reason to keep them.  We obviously can't stop at only having Flame type boosters, so to complete the entire set we'd have to make an ability for each effect (2), each stat (3), and each element (19) for a total of 144 abilities that basically do the same thing.  Not worth it in my opinion.  As such they're being removed and save files will be patched, once again.  Sorry, but I promise it's for the best.