Updated game with ability sprites, parameters, and functions for Slash Claw, added to ability generator.
Slash Claw has been re-classed as a pure Cutter type and a new placeholder for Slash Man's second tier ability has been added with the new Slash Beast (pending).  Slaw Claw's ability type not matching that of it's owner's core is a first for the RPG, but hopefully not the last as it put him in a very unique position ripe for new strategies.