Updated Star Crash with new mechanics based on feedback from the community.
Star Crash is now an offensive-only ability thrown at a target for up to three times damages as it spins.  The Shield typing is now reflective only of its appearance/structure and not necessarily its function, while the Space typing should be obvious.  The previous damage-absorbing functionality has been backed up and copied to Junk Shield with a few tweaks so it's ready when the sprite is.