Updated robot editor with some of the visual changes imported from MMRPG World.
This includes the core icon in the header, the robot sprite showing under the mugshot, the summon count under the robot sprite, and the life and weapon energy being shown side-by-side and always in colour without the redundant calculation text that was there before.  Robot editor updates will continue until I've got it working as fast and smoothly as the World version.