Updated profile editor with the ability to change your password, added new omega seed.
A user's *first* encoded password will now be used as their "omega seed" going forward.  This seed will be used for naming their save file and possibly for future procedurally generated content.  All other parts of the website and game have been updated to respect this change, and to no longer store the actual password in the session.  Lastly, the save file itself now contains much less data and exists primary for legacy purposes - everything else is stored in the database.  The folder may be used for user-generated-content at a later date, just not now.