Updated shop files with new "alts" section for Auto's shop, allowing the users to purchase alt outfits instead of unlocking via summon.
Auto's Shop expands to include alts at Level 40 and continues expanding with more alts every 5 levels.  The price of alts is directly based on summon count (summon count * 1000) with a large discount applied for all of the support bots (summon count * 100).  Alts unlocked in this way can be used in all the same places as alts unlocked via summon.