Updated the swap abilities to function differently for support robots vs other.
Support robots are still able to swap stats with the enemy, but when used by any other robot these abilities will swap with a teammate instead.  This is technically a nerf, but it also provides a lot of new strategies so I think it's the right decision.  Next up, the shuffle abilities...