Added new "Energy Mode" ability to replace Repair Mode.
Ability is identical to old one but renaming would have been too annoying to write a patch for.  Instead, Repair Mode was cloned into this new ability and the old one will simply be "hidden" in the game.  Anyone with the old ability can keep it, but it really makes no difference.