Updated single and double mission generators to occasionally attach hold items to enemy robot masters.
Various stat pellets (single missions) and capsules (double missions) as well as cores (doubles only) may be attached, but only the cores have an effect at this time granting the holder compatibility with more ability types.  The plan is to have consumable items like the Energy Pellet auto-trigger like a berry would in Pokémon, but it's not yet programmed in.