Updated robot editor to load a player canvas in addition to abilities and items.
With the dropdown for players no longer working on mobile, we need to update player transfer UI to work the same as abilities and items.  This patch adds functions for printing player data in the editor as well as script and style changes to load and display the option links.  No other functionality has been programed, however, as that will come in the next commit.