Updated the in-battle scale and positioning of sprites with new perspective mode.
Allowed users to go back to the old mode by toggling "Perspective Mode" from the in-battle menu.  This new mode positions sprites in a way that makes more sense, with the active robot in the middle-front of its benched allies instead of them appearing in a line.  This means that 1on1 battles will be zoomed in like before but battles with 2 - 8 opponents will appear more zoomed out then before.