Updated item class with support for three more description field types.
Now description text can be split into distinct parts for basic info, use text, hold text, and price text.  This way we can display only what's relevant in various contexts of the game.  The Small and Large Screws have been updated with this new format and the others are forthcoming.