Created the new Core Shield and Core Laser abilities, fleshed out their functionality.
These abilities are single-type (Shield and Laser respectively) but each take on a secondary element if the holder is a non-Neutral core robot or is holding a non-Neutral core item.  In the case that both of these factors are true, the held item's type takes precedence in determining the secondary element.  This ability will remain single-type if primary and secondary types are the same.

The Core Shield ability cloaks the user in an elemental barrier that protects them from 99% of all damage that matches the shield's type.  If the user is holding a Target Module, the shield can be attached to benched teammates instead.

The Core Laser ability fires a powerful beam of energy across the field to deal damage at a single target.  If the user is holding a Target Module, the laser can be directed at a benched opponent instead.