Allowed users and admins to edit their omega sequence, added salt values to all website encryption for security.
Also removed references and requirements for a literal save "file" to be created before logging in and saving the game.  Folders for each user can be created when it's actually necessary.  This change required many updates across any and all files relating to user/game values so there may be bugs.