Updated session-to-database function to skip players or robots that are not valid (anymore).
Old entries for the different generations of mecha, for example, no longer exist and were inflating the database.  Additionally, some characters like "player" and "robot" were erroneously unlocked for some users and needed to be skipped as well.  Will probably add another flag at some point in the future for robots that are specifically unlockable so we can exclude restricted ones more easily.