Added new function for calculating max robot limit for the player.
Going forward, players may only bring 2 robots into battle for each player they've unlocked plus 2 more once they've completed the prototype as all three (for a total of 8 robots in battle).  This forces a bit of strategy as bringing 8vs8 from the get-go is too easy and needs to be corrected.