Refined item dropping mechanics, fixed bugs + improved functionality
Firstly, fixed a bug that made it possible to gain more than 4 shards and trigger an overflow error.  Next, made it so the game doesn't bother including items the user already has max quantity of (4 for shards and 99 for everthing else). This ensures screws are awarded instead of shards/cores when the user is already full of the latter.  Lastly, made it so that shards/cores *replace* all other items instead of adding-on to them when applicable.  This ensures the item dropped is actually a core/shard when you manage to trigger a weakness and not something you didn't intend.