Updated Copy Sould to generate/transform held cores where applicable
If the user is not holding an item, a new core will be generated.  If the user is already holding a core, that core\'s type will be changed.  If the user is holding any other type of item, then the user's own core is changed to the copied type instead.