Improved mechanics for held cores, core shields, and copy soul
Holding an elemental core now auto-casts a matching Core Shield on the user at the beginning of the battle.  Using the Copy Soul ability also auto-casts a matching Core Shield where applicable.  The duration of Core Shield depends on how it was created; 3 when done via Copy Soul ability, 6 when done via hold item, and 9 when done via the Core Shield ability.