Removed mecha shards from BP calculations
Their mechanics don't lend themselves to being collectible in the same way the other items are.  It would be like The Legend of Zelda requiring you to collect an additional 4 Pieces of Heart after you've already completed all your Heart Containers.  Sorry for the loss of BP everyone, but it's a nerf being applied equally so in relative terms nothing is really happening changing.