Added data + images for the Dynamo Bolt ability
Based on the original 'Lightning Bolt' but tweaked for the RPG, the move heals on the first turn and then releases an AoE lightning attack on the second.  Outgoing damage from this ability is reduced if the user is at low health, with the user needing >50% to deal anything higher than 1 point.