Tied bonus chapter unlocks to doctor program items
Player Battles are unlocked via the Light Program, Challenge Mode is unlocked via the Wily Program, and Star Fields are unlocked via the Cossack Program. The updated descriptions will tie into the lore eventually, but for right now the changes are mainly to improve the flow of the game and its unlockables.