Added new Challenge Board to player editor
Challenge board shows available challenges and players decide which ones they want in each doctor's campaign. This will be important going forward now that challenges contribute to BP and leaderboard rankings. At the moment only event-type challenges are available to normal users (all for admins) but we'll have a way for user-generated challenges to be selected in the future.