Revised Crystal Eye functionality and animation
Ability still splits and hits benched robots, BUT the ability type is based on the target only and hits the benched robots with the same element (so only the active should be super effective).  Additionally, instead of hitting the two ends of the bench, it specifically hits the two spaces behind the target if robots are in them.