Updated mechanics for Robot Cores & Core Shields
Core Shields now last for 3 turns regardless of how they were summoned. Period. When a robot is HOLDING a core, however, that shield gets a +1 duration at the end of each turn.  This means that as long as a robot continues holding a core, they will have a shield.  If the core is dropped/removed/swapped, then the shield will last for 3 turns before disappearing.  Because it's possible for multiple shields to exist on a robot at once, trading held cores mid-battle is a good way to stack protections quickly.