Revised Energy Swap ability with cleaner mechanics
Now you swap current energy amounts only, and if cropping occurs then so be it.  If you have 300/300 and the target has 100/200 but uses Energy Swap on you, then you will be left with 100/300 and they'll be at 200/200 or full health. The code is cleaner, it doesn't interfere with the Energy Upgrade item, and it can be used and defended against more strategically.