Fixed endless mode bug w/ only mecha remains
If the user ended an Endless Mode battle with only a summoned mecha remaining, the game would automatically try to resume the Endless Mode run but be unable to due to the lack of a proper robot master, making the game 100% impossible to play.  Now it catches that instance and redirects them back to home, ending the run automatically.