Updated rpg_ability.php w/ hotfix for Buster Shot's max-shot limit.
So I guess last year when we were balancing the Buster Shot (or rather, tweaking it to prevent server-overload from bad-math), we made it so that the max-shots were limited by the user's current level.  Maybe this was an easy to way "limit it by 100" without explicitly saying so(?) but it makes the early game confusing because the Buster Shot isn't hitting the Met as many times as you expect in the opening battle + others and it doesn't tell you *why* anywhere in the description.  I considered amending the description, but I'd rather just fix the oversight.  So I made the Buster Shot's max-shots limited to x99 unless the user is >= level 100, in which case it uses the user's level directly.  Should have done this a while ago.  Updates to rpg_ability.php and buster-shot/functions.php will use the same code changes FWIW.